move handleSelectChange from iam to lib, and add unit tests
